pyproject.toml
[project]
name = "cctv-uploader"
version = "0.1.0"
description = "CCTV video uploader service"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "boto3",
    "requests",
    "python-dotenv",
    "sqlalchemy",
    "psycopg2-binary",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
